refactor(console): split api endpoints in modules
