+++ src/console/src/api/invitation_code.rs
@@ -1,0 +1,9 @@
+use crate::store::heap::add_invitation_code as add_invitation_code_store;
+use crate::types::state::InvitationCode;
+use ic_cdk_macros::update;
+use crate::guards::caller_is_admin_controller;
+
+#[update(guard = "caller_is_admin_controller")]
+fn add_invitation_code(code: InvitationCode) {
+    add_invitation_code_store(&code);
+}
