+++ src/console/src/api/mod.rs
@@ -1,0 +1,9 @@
+mod cdn;
+mod controllers;
+mod economy;
+mod factory;
+mod http;
+mod invitation_code;
+mod mission_control;
+mod rates;
+mod storage;
